Add files via upload
In previous version the random.shuffle was done on numpy array, which apparently wasnt really bright of me, so now its done on the training data before the numpy array is made out of it. Also, it is now possible to get the weights and biasis from the training when it was at the lowest error either for training data or for test data. Lastly, the weights and biases are initialized as in already somewhat trained network, altough that can be changed easily.