Add files via upload
The way of getting the nice weights and biases was better in the version before the previous one. Here that same way of writing them is used for erwb (the rest remains the same as previous version). Most importantly, the weights and biases that it initializes, are in this new version a bit better then in the previous version (avarage error is now roughly 0.1).